fixed error

+++ src/main/java/mbti/restapi/BoardAPI.java
@@ -52,7 +52,14 @@ public class BoardAPI {
 	}
 
 	@DeleteMapping("board/{boardSeq}")
-	public boolean delete(@PathVariable(name = "boardSeq") int boardSeq){
+	public boolean delete(@PathVariable(name = "boardSeq") int boardSeq, HttpSession session){
+		if (session.getAttribute("member") == null) {
+			return false;
+		}
+		int writerSeq = boardService.findMemberSeqByBoardSeq(boardSeq);
+		if (writerSeq!=Integer.parseInt(session.getAttribute("memberSeq").toString())) {
+			return false;
+		}
 		boardService.delete(boardSeq);
 		return true;
 	}
